backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-bcrypt==1.0.1
flask-jwt-extended==4.5.3
pymongo==4.6.1
python-dotenv==1.0.0
PyJWT==2.8.0
